+++ packages/ui-components/src/postgrest-infinite-scroll.ts
@@ -3,7 +3,13 @@ import { InfiniteScrollProps, InfiniteScrollView } from "./infinite-scroll";
 import { useAPIResult } from "./api";
 import { useMemo, useState } from "react";
 import { MultiSelect, ItemRenderer, ItemPredicate } from "@blueprintjs/select";
-import { MenuItem, Spinner, InputGroup } from "@blueprintjs/core";
+import {
+  MenuItem,
+  Spinner,
+  InputGroup,
+  Collapse,
+  Icon,
+} from "@blueprintjs/core";
 import styles from "./postgrest.module.sass";
 
 const h = hyper.styled(styles);
@@ -11,6 +17,7 @@ const h = hyper.styled(styles);
 interface PostgRESTInfiniteScrollProps extends InfiniteScrollProps<any> {
   id_key: string;
   limit: number;
+  filter_threshold?: number;
   extraParams?: Record<string, any>;
   ascending?: boolean;
   filterable?: boolean;
@@ -32,6 +39,12 @@ interface PostgRESTInfiniteScrollProps extends InfiniteScrollProps<any> {
     tagInputProps: any;
     popoverProps: any;
   }>;
+  group_key?: string;
+  groups?: Array<{ value: string; label: string }>;
+  GroupingComponent?: React.ComponentType<{
+    title: string;
+    children: React.ReactNode;
+  }>;
 }
 
 export function PostgRESTInfiniteScrollView(
@@ -52,8 +65,11 @@ export function PostgRESTInfiniteScrollView(
     MultiSelectComponent,
     extraParams = {},
     key,
+    GroupingComponent,
     toggles = null,
     searchColumns = undefined,
+    group_key = undefined,
+    filter_threshold = 0,
     ...rest
   } = props;
 
@@ -66,6 +82,7 @@ export function PostgRESTInfiniteScrollView(
   );
 
   const [filterValue, setFilterValue] = useState<string>("");
+  const hideData = filterValue.length < filter_threshold;
 
   const SearchBarToUse = SearchBarComponent ?? SearchBar;
   const MultiSelectToUse = MultiSelectComponent ?? MultiSelect;
@@ -236,35 +253,53 @@ export function PostgRESTInfiniteScrollView(
         h(SearchBarToUse, {
           onChange: (value) => setFilterValue(value || ""),
         }),
-        h(MultiSelectToUse, {
-          items: keys.filter((item) => !selectedItems.includes(item.value)),
-          itemRenderer,
-          itemPredicate: filterItem,
-          selectedItems,
-          onItemSelect: handleSelect,
-          tagRenderer: (value) => {
-            const found = keys.find((k) => k.value === value);
-            return found ? found.label : value;
-          },
-          onRemove: handleRemove,
-          tagInputProps: {
+        h.if(searchColumns == null || searchColumns.length > 1)(
+          MultiSelectToUse,
+          {
+            items: keys.filter((item) => !selectedItems.includes(item.value)),
+            itemRenderer,
+            itemPredicate: filterItem,
+            selectedItems,
+            onItemSelect: handleSelect,
+            tagRenderer: (value) => {
+              const found = keys.find((k) => k.value === value);
+              return found ? found.label : value;
+            },
             onRemove: handleRemove,
-            placeholder: "Select a column(s) to filter by...",
+            tagInputProps: {
+              onRemove: handleRemove,
+              placeholder: "Select a column(s) to filter by...",
+            },
+            popoverProps: { minimal: true },
           },
-          popoverProps: { minimal: true },
-        }),
+        ),
       ]),
       h.if(toggles)("div.toggles", toggles),
     ]),
-    h(InfiniteScrollView, {
-      ...rest,
-      route,
-      getNextParams: getNextParams ?? defaultGetNextParams,
-      params: params ?? defaultParams,
-      initialItems: newInitialItems,
-      hasMore: hasMore ?? defaultHasMore,
-      key: newKey,
-    }),
+    group_key
+      ? Grouping({
+          group_key,
+          groups: props.groups ?? [],
+          route,
+          id_key,
+          params: defaultParams,
+          getNextParams: getNextParams ?? defaultGetNextParams,
+          hasMore: hasMore ?? defaultHasMore,
+          key: newKey,
+          rest,
+          hideData,
+          GroupingComponent,
+        })
+      : h(InfiniteScrollView, {
+          ...rest,
+          route,
+          getNextParams: getNextParams ?? defaultGetNextParams,
+          params: params ?? defaultParams,
+          initialItems: newInitialItems,
+          hasMore: hasMore ?? defaultHasMore,
+          key: newKey,
+          hideData,
+        }),
   ]);
 }
 
@@ -281,4 +316,120 @@ function SearchBar({ onChange, placeholder = "Search..." }) {
     },
     leftIcon: "search",
   });
+}
+
+interface GroupingProps {
+  group_key: string;
+  groups: Array<{ value: string; label: string }>;
+  route: string;
+  id_key: string;
+  params?: Record<string, any>;
+  getNextParams?: (
+    response: any[],
+    params: Record<string, any>,
+  ) => Record<string, any>;
+  hasMore?: (response: any[]) => boolean;
+  key?: string;
+  rest?: any;
+  hideData?: boolean;
+  GroupingComponent?: React.ComponentType<GroupingProps>;
+}
+
+function Grouping(props: GroupingProps) {
+  const {
+    group_key,
+    groups,
+    route,
+    id_key,
+    params,
+    getNextParams,
+    hasMore,
+    rest,
+    hideData,
+    GroupingComponent,
+  } = props;
+
+  return h("div.group-page", [
+    groups.map((group) => {
+      if (!group.value || !group.label) {
+        throw new Error("Each group must have a value and label");
+      }
+
+      return h(GroupPanel, {
+        group,
+        route,
+        id_key,
+        hideData,
+        params: {
+          ...params,
+          [group_key]: "eq." + group.value,
+        },
+        getNextParams,
+        hasMore,
+        ...rest,
+        GroupingComponent,
+      });
+    }),
+  ]);
+}
+
+function GroupPanel(props) {
+  const {
+    group,
+    route,
+    params,
+    getNextParams,
+    hasMore,
+    key,
+    hideData,
+    GroupingComponent,
+    ...rest
+  } = props;
+
+  const data = useAPIResult(route, {
+    ...params,
+    limit: 1,
+  });
+
+  if (!data || data?.length === 0 || hideData) return null;
+
+  const Panel = GroupingComponent || ExpansionPanel;
+
+  return h(
+    Panel,
+    {
+      title: group.label,
+    },
+    [
+      h(InfiniteScrollView, {
+        key: key || group.value,
+        route,
+        params,
+        getNextParams,
+        hasMore,
+        hideData,
+        ...rest,
+      }),
+    ],
+  );
+}
+
+function ExpansionPanel(props) {
+  const { title, children, className } = props;
+  const [isOpen, setOpen] = useState(false);
+
+  return h(
+    "div.expansion-panel",
+    { className: `expansion-panel ${className || ""}` },
+    [
+      h("div.panel-header", { onClick: () => setOpen(!isOpen) }, [
+        h("h2.title", title),
+        h(Icon, {
+          icon: isOpen ? "chevron-up" : "chevron-down",
+          className: "expansion-panel-icon",
+        }),
+      ]),
+      h(Collapse, { isOpen }, h("div.child-container", children)),
+    ],
+  );
 }